+++ packages/mobile/src/components/core/index.ts
@@ -37,3 +37,4 @@ export * from './ProfilePicture'
 export * from './UserCoverPhoto'
 export * from './UserDisplayName'
 export * from './DateTimeInput'
+export * from './ExpandableContent'
